vim-patch:8.1.1743: 'hlsearch' and match highlighting in the wrong place
Problem:  'hlsearch' and match highlighting in the wrong place.
Solution: Move highlighting from inside screen functions to highlight.c.

https://github.com/vim/vim/commit/bbca7732e8a3deb6e5dcf84739579a2667a75475
